Fix GCC warnings

+++ src/Headers/bg3d.h
@@ -13,7 +13,7 @@
 												
 #define	MAX_BG3D_MATERIALS			400			// max # of materials in a bg3d file
 
-#define	MAX_BG3D_GROUPS			(MODEL_GROUP_SKELETONBASE+MAX_SKELETON_TYPES)	// skeletons are @ end of list, so can use these counts for max #
+#define	MAX_BG3D_GROUPS			((int)MODEL_GROUP_SKELETONBASE+(int)MAX_SKELETON_TYPES)	// skeletons are @ end of list, so can use these counts for max #
 #define	MAX_OBJECTS_IN_GROUP	100
 
 		/***********************/
